requirements.txt
anthropic==0.7.4
flask==3.0.0
tavily-python==0.3.1
python-dotenv==1.0.0